use coarsetime::Instant::recent() instead
