+++ static-metric/examples/make_auto_flush_static_metric_histogram.rs
@@ -11,12 +11,9 @@ extern crate coarsetime;
 extern crate prometheus;
 extern crate prometheus_static_metric;
 
-use coarsetime::Updater;
 use prometheus::*;
 use prometheus_static_metric::auto_flush_from;
 use prometheus_static_metric::make_auto_flush_static_metric;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
 
 make_auto_flush_static_metric! {
 
@@ -56,18 +53,7 @@ lazy_static! {
         auto_flush_from!(HTTP_HISTO_VEC, Lhrs, std::time::Duration::from_secs(1));
 }
 
-lazy_static! {
-    static ref UPDATER_IS_RUNNING: Arc<AtomicBool> = Arc::new(AtomicBool::new(false));
-}
-
-fn ensure_updater() {
-    if !UPDATER_IS_RUNNING.compare_and_swap(false, true, Ordering::SeqCst) {
-        Updater::new(200).start().unwrap();
-    }
-}
-
 fn main() {
-    ensure_updater();
     TLS_HTTP_COUNTER.foo.post.http1.observe(1.0);
     TLS_HTTP_COUNTER.foo.post.http1.observe(1.0);
